pyproject.toml
[tool.poetry]
name = "roundagain"
version = "0.1.0"
description = "A connection management application."
authors = ["Eric Dudley <17283326+ericdudley@users.noreply.github.com>"]
license = "GPL-3.0-only"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.13"
flask = ">=3.1.0,<4.0.0"
sqlalchemy = ">=2.0.40,<3.0.0"
apscheduler = ">=3.11.0,<4.0.0"
htmx-flask = ">=0.1.0,<0.2.0"
flask-wtf = ">=1.2.2,<2.0.0"
gunicorn = ">=23.0.0,<24.0.0"
python-dotenv = ">=1.1.0,<2.0.0"
jinja2 = ">=3.1.6,<4.0.0"
email-validator = ">=2.2.0,<3.0.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=8.3.5"
pytest-flask = ">=1.3.0"
black = ">=25.1.0"
flake8 = ">=7.2.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"